+++ global.R
@@ -1,13 +1,14 @@
-# Install PoolTestR: r-universe dev version
+# # Install PoolTestR: r-universe dev version
+# # To list repos in current session: getOption("repos")
 # options(repos = c(
 #   CRAN = "https://cloud.r-project.org",
 #   angusmclure = "https://angusmclure.r-universe.dev"
 # ))
-# install.packages("PoolTestR")
+# install.packages("PoolTestR", repos = c("https://angusmclure.r-universe.dev"))
 
 # Install PoolPoweR: github dev version
 # devtools::install_github("AngusMcLure/PoolPoweR")
-appVersion <- "0.1.4"
+appVersion <- "0.1.5"
 
 library(shiny)
 library(shinyBS)